+++ package.json
@@ -11,6 +11,7 @@
     "@material-ui/core": "^5.0.0-beta.5",
     "@material-ui/lab": "^5.0.0-alpha.44",
     "@material-ui/styled-engine-sc": "^5.0.0-beta.5",
+    "@material-ui/styles": "^4.11.4",
     "@mui/icons-material": "^5.0.5",
     "@mui/lab": "^5.0.0-alpha.53",
     "@mui/material": "^5.0.4",
