+++ src/pages/myPage/information/Information.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Information() {
+  return <div></div>;
+}
